+++ src/shared/event/client_trigger.rs
@@ -12,7 +12,9 @@ use super::{
     remote_event_registry::RemoteEventRegistry,
     remote_targets::RemoteTargets,
 };
-use crate::shared::{backend::replicon_channels::Channel, entity_serde, postcard_utils};
+use crate::shared::{
+    backend::replicon_channels::Channel, entity_serde, postcard_utils, protocol::ProtocolHasher,
+};
 
 /// An extension trait for [`App`] for creating client triggers.
 ///
@@ -74,6 +76,10 @@ impl ClientTriggerAppExt for App {
         deserialize: EventDeserializeFn<ServerReceiveCtx, E>,
     ) -> &mut Self {
         debug!("registering trigger `{}`", any::type_name::<E>());
+
+        self.world_mut()
+            .resource_mut::<ProtocolHasher>()
+            .add_client_trigger::<E>();
 
         let event_fns = EventFns::new(serialize, deserialize)
             .with_outer(trigger_serialize, trigger_deserialize);